+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_extracting_callbacks/config_parser.dart
@@ -1,0 +1,11 @@
+part of 'prefer_extracting_callbacks.dart';
+
+class _ConfigParser {
+  static const _ignoredArgumentsConfig = 'ignored-named-arguments';
+
+  static Iterable<String> parseIgnoredArguments(Map<String, Object> config) =>
+      config.containsKey(_ignoredArgumentsConfig) &&
+              config[_ignoredArgumentsConfig] is Iterable
+          ? List<String>.from(config[_ignoredArgumentsConfig] as Iterable)
+          : <String>[];
+}